% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{as.data.frame.Dataset}
\alias{as.data.frame.Dataset}
\title{Create a data.frame object from a Dataset.}
\usage{
\method{as.data.frame}{Dataset}(x, ...)
}
\arguments{
\item{x}{A Dataset object.}
}
\value{
A data.frame with the actual data of the original Dataset.
}
\description{
This function returns the actual data of a Dataset as a data.frame.
Additional information associated with a Dataset are neglected.
}
