% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{configuration<-}
\alias{configuration<-}
\title{Configuration function for changing config of an Inducer object}
\usage{
configuration(inducer) <- value
}
\arguments{
\item{inducer}{An inducer being an Inducer object.}

\item{value}{the new value for the config in Inducer}
}
\description{
change values of the configuration of an Inducer.
}
