% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{configuration.Inducer}
\alias{configuration.Inducer}
\title{S3 method configuration for class \code{inducer}}
\usage{
\method{configuration}{Inducer}(.inducer, ...)
}
\arguments{
\item{.inducer}{an Inducer object}

\item{...}{further arguments}
}
\description{
Get the configuration of an \code{inducer} object
}
\examples{
inducer <- InducerXGBoost()
configuration(inducer)
}
