% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{configuration.Models}
\alias{configuration.Models}
\title{Configuration method for Models objects}
\usage{
configuration.Models(model, ...)
}
\arguments{
\item{model}{A Models object for which the configuration should be obtained.}
}
\value{
The configuration of the model.
}
\description{
Get the configuration of a Models object.
}
