% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{configuration}
\alias{configuration}
\title{Get the configuration of an inducer}
\usage{
configuration(inducer)
}
\arguments{
\item{inducer}{An Inducer object for which the hyperparameter configuration
should be obtained.}
}
\value{
The hyperparameter configuration of a given inducer.
}
\description{
Get the hyperparameter configuration of an inducer.
}
