% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Dataset}
\alias{Dataset}
\title{Create a \code{Dataset} object}
\usage{
Dataset(
  data,
  target,
  type = NULL,
  name = as.name(deparse(substitute(data), 20)[[1]])
)
}
\arguments{
\item{data}{A matrix or data.frame object with relevant data and named columns.}

\item{target}{A string of a column name of data specifying the target.}

\item{type}{A string specifying whether a regression or classification should be done.}

\item{name}{Name attribute that will be stored internally.}
}
\value{
An object of class 'Dataset' with attributes 'data' containing the actual data as a data.frame,
'target' with the name of the target covariable, 'type' which is either 'classification' or 'regression'
and 'name'.
}
\description{
From a given named \code{matrix} or a named \code{data.frame} a \code{Dataset} is created.
The user needs to specify the target covariable/columnname as a string.
Optionally the \code{type} can be set to either \code{regression} or \code{classification}.
If this is not provided the target will be automatically inferred from
the type in matrix or data.frame.
}
\examples{
cars.data <- Dataset(cars, target = "dist")

}
