% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorAIC.R
\name{EvaluatorAIC}
\alias{EvaluatorAIC}
\title{Evaluate performance through AIC}
\usage{
EvaluatorAIC(.model)
}
\arguments{
\item{.model}{A \code{ModelLm} object for which the performance should be measured.}
}
\value{
The AIC value of the fitted model.
}
\description{
Evaluate the performance of a \code{ModelLm} using the Akaike information criterion.
}
\examples{
cars.ds <- Dataset(data = dataset::cars, target = "dist")
mod_fit <- fit(InducerLm(), cars.ds, formula = as.formula("dist~speed"))
EvaluatorAIC(mod_fit)
}
\seealso{
\code{\link[=EvaluatorBIC]{EvaluatorBIC()}} to get the Bayesian information criterion (BIC), \code{\link[=EvaluatorMAE]{EvaluatorMAE()}} for computing the
mean absolute error, \code{\link[=EvaluatorMSE]{EvaluatorMSE()}} for the mean-squared error.
}
