% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorAccuracyROC.R
\name{EvaluatorAUC}
\alias{EvaluatorAUC}
\title{Compute Area Under The Curve}
\usage{
EvaluatorAUC(.prediction, .dataset, .target)
}
\arguments{
\item{.prediction}{A \code{data.frame} containing the predictions and the true values as columns or a
numeric vector containing only the predictions. The true values have to be encoded by 0/1
or by TRUE/FALSE. The predicted values have to be numeric and be in a range of 0 to 1.}

\item{.dataset}{An optional \code{Dataset} or \code{data.frame} object, that has to be provided if \code{.prediction} is
a numeric vector.}

\item{.target}{A character vector of length one, being the name of the target variable contained as column
in the .dataset}
}
\description{
Evaluate the performance of a 'ModelClassification' object on a binary classification problem
using the area under the ROC curve.
}
\examples{
x <- data.frame(var1 = c(1, 2, 3, 4, 5, 6, 7), target = c(1, 1, 1, 1, 0, 1, 0))
predictions <- c(1)
EvaluatorAUC(predictions, x, "target")
predictions <- data.frame(prediction = c(0.8, 0.2, 0.6, 0.8, 0.8), truth = c(1, 0, 1, 1, 1))
EvaluatorAUC(predictions)
predictions <- data.frame(prediction = c(0.8, 0.2, 0.6, 0.8, 0.8), truth = c(TRUE, FALSE, TRUE, TRUE, TRUE))
EvaluatorAUC(predictions)
}
\seealso{
\code{\link[=EvaluatorAccuracy]{EvaluatorAccuracy()}} for evaluating the accuracy of a classifier, \code{\link[=EvaluatorMAE]{EvaluatorMAE()}} for computing the
mean absolute error, \code{\link[=EvaluatorMSE]{EvaluatorMSE()}} for the mean-squared error (corresponding to the Brier-Score in binary
classification).
}
