% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorAIC.R
\name{EvaluatorBIC}
\alias{EvaluatorBIC}
\title{Evaluate performance through BIC}
\usage{
EvaluatorBIC(.model)
}
\arguments{
\item{.model}{A \code{ModelLm} object for which the performance should be measured.}
}
\value{
The BIC value of the fitted model.
}
\description{
Evaluate the performance of a \code{ModelLm} using the Bayesian information criterion.
}
\examples{
cars.ds <- Dataset(data = cars, target = "dist")
mod_fit <- fit(InducerLm(), cars.ds, formula = as.formula("dist~speed"))
EvaluatorBIC(mod_fit)
}
\seealso{
\code{\link[=EvaluatorAIC]{EvaluatorAIC()}} to get Akaike information criterion (AIC), \code{\link[=EvaluatorMAE]{EvaluatorMAE()}} for computing the
mean absolute error, \code{\link[=EvaluatorMSE]{EvaluatorMSE()}} for the mean-squared error.
}
