% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorMAE_MSE.R
\name{EvaluatorMSE}
\alias{EvaluatorMSE}
\title{Evaluate predictions using the Mean Squared Error}
\usage{
EvaluatorMSE(.prediction, .data, .target)
}
\arguments{
\item{.prediction}{A \code{data.frame} object containing the predictions of a model. The columns should contain
the predictions and the true values. If only the predictions are handed over, the true values of the target
variable have to be handed over in the \code{.data} and the name of the target variable in the \code{.target} argument.}

\item{.data}{Optional argument, which has to be provided if only the predictions are handed over in \code{.prediction},
\code{.data} has to be a \code{data.frame} which contains the true values.}

\item{.target}{If only the predictions are handed over in \code{.prediction}, \code{.target} has to be handed over as
a \code{character}.
of length 1 being the name of the target variable.}
}
\value{
The mean squared error of the predictions.
}
\description{
Compute the mean squared error of the predictions of a model. For binary classification this
corresponds to the Brier-Score.
}
\examples{
x <- data.frame(var1 = c(1, 1, 1, 1, 0), target = c(1, 2, 3, 4, 5))
predictions <- c(3)
EvaluatorMSE(predictions, x, "target")
predictions <- data.frame(prediction = c(1.3, 2.5, 2.6, 3.5, 4.5), truth = c(1, 2, 3, 4, 5))
EvaluatorMSE(predictions)
}
\seealso{
For further Evaluators: \code{\link[=EvaluatorMAE]{EvaluatorMAE()}}, \code{\link[=EvaluatorAIC]{EvaluatorAIC()}}, \code{\link[=EvaluatorBIC]{EvaluatorBIC()}},
\code{\link[=EvaluatorAccuracy]{EvaluatorAccuracy()}}, \code{\link[=EvaluatorAUC]{EvaluatorAUC()}}
}
