% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_ModelLm.R
\name{fit.InducerLm}
\alias{fit.InducerLm}
\title{Fit a Model using \code{InducerLm}}
\usage{
\method{fit}{InducerLm}(
  .inducer,
  .data,
  formula,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  offset,
  ...
)
}
\arguments{
\item{.inducer}{An \code{InducerLm} object. The Inducer which should be used for the fitting.}

\item{.data}{The data to which the model should be fitted, provided as a \code{Dataset} object.}

\item{formula}{An object of class \code{formula}.
An optional parameter setting the \code{formula} argument of an \code{InducerLm} object.}

\item{subset}{An optional argument. A vector specifying a subset of observations that should be used
for fitting the model.}

\item{weights}{An optional argument. A vector of weights that should be used for fitting the model.}

\item{na.action}{An optional argument. A function that specifies how to handle missing values.}

\item{method}{The method which should be used for fitting. For more information see \link{lm}}

\item{model}{if true model is returned}

\item{x}{if true x is returned}

\item{y}{if true y is returned}

\item{qr}{if true the QR decomposition is returned}

\item{singular.ok}{logical. If FALSE (the default in S but not in R) a singular fit is an error}

\item{offset}{this can be used to specify an a priori known component to be included in the
linear predictor during fitting.}
}
\value{
An object of class \code{ModelLm}.
}
\description{
Fit a linear model on the provided data.
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
inducer <- InducerLm()
lmfit <- fit.InducerLm(.inducer = inducer, .data = cars.data)
}
