% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_ModelRpart.R
\name{fit.InducerRpart}
\alias{fit.InducerRpart}
\title{Fit a Model using \code{InducerRpart}}
\usage{
\method{fit}{InducerRpart}(
  .inducer,
  .data,
  formula,
  weights,
  subset,
  na.action = "na.rpart",
  method,
  model = FALSE,
  x = FALSE,
  y = TRUE,
  parms,
  control,
  cost,
  ...
)
}
\arguments{
\item{.inducer}{An \code{InducerRpart} object. The Inducer which should be used for the fitting.}

\item{formula}{An optional parameter setting the \code{formula} argument of an \code{InducerRpart} object.}

\item{weights}{optional case weights.}

\item{subset}{optional expression saying that only a subset of the rows of the data should be used in the fit.}

\item{na.action}{the default action deletes all observations for which y is missing, but keeps those in
which one or more predictors are missing.}

\item{method}{one of "anova", "poisson", "class" or "exp". If method is missing then the routine tries
to make an intelligent guess}

\item{model}{if logical: keep a copy of the model frame in the result? If the input value for model is a model frame}

\item{x}{keep a copy of the x matrix in the result.}

\item{y}{keep a copy of the dependent variable in the result.
If missing and model is supplied this defaults to FALSE.}

\item{parms}{optional parameters for the splitting function.}

\item{control}{a list of options that control details of the rpart algorithm}

\item{cost}{a vector of non-negative costs, one for each variable in the model. Defaults to one for all variables}

\item{data}{The data to which the model should be fitted, provided as a \code{Dataset} object.}
}
\value{
An object of class \code{InducerRpart}.
}
\description{
Fit a rpart model on the provided data.
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
inducer <- InducerRpart()
lmfit <- fit.InducerRpart(.inducer = inducer, .data = cars.data)
}
