% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Models.R
\name{fit}
\alias{fit}
\title{S3 method fit}
\usage{
fit(...)

fit(...)
}
\arguments{
\item{...}{further arguments}

\item{x}{An inducer which should be used for the model fitting.}
}
\value{
A fitted model.
}
\description{
\code{fit} fits a model on given data.

fit a model.
}
