% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter.R
\name{hp}
\alias{hp}
\title{Hyperparameter
This function return the name, type and range of hyperparameters}
\usage{
hp(...)
}
\arguments{
\item{...}{optional arguments}
}
\value{
This function return the name, type and range of hyperparameters
}
\description{
Hyperparameter
This function return the name, type and range of hyperparameters
}
