% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter.R
\name{hyperparameter}
\alias{hyperparameter}
\title{Hyperparameter
This function return the name, type and range of hyperparameters}
\usage{
hyperparameter(...)
}
\description{
Hyperparameter
This function return the name, type and range of hyperparameters
}
\examples{
# hpx <- hp(x = p_num(0, 1), y = p_int(1, Inf), z = p_fct(letters))
}
