% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{hyperparameters}
\alias{hyperparameters}
\title{Get hyperparameters of an inducer}
\usage{
hyperparameters(inducer)
}
\arguments{
\item{inducer}{An Inducer object for which the hyperparameters should
be obtained.}
}
\value{
The hyperparameters of the given inducer and their range.
}
\description{
Get the hyperparameters of an inducer.
}
