% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{inducer.Models}
\alias{inducer.Models}
\title{inducer: get the inducer and its configuration}
\usage{
inducer.Models(model, ...)
}
\arguments{
\item{model}{an object of class Models, for which the inducer and the configuration should be returned.}
}
\value{
The inducer and the configuration used for the model.
}
\description{
This function returns the inducer and the hyperparameter setting used for the model fitting.
}
\examples{
TODO!!!!

}
