% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{Inducer}
\alias{Inducer}
\title{Create an Inducer}
\usage{
Inducer(.data = NULL, name, configuration, defaults, hyperparameter)
}
\arguments{
\item{name}{The name of the inducer}

\item{configuration}{The configuration of the inducer.}

\item{defaults}{The default values of the inducer.}

\item{hyperparameter}{A named list containing the hyperparameters of the inducer.
The list should contain the names of the hyperparameters, the type of the hyperparameters,
the lower and upper bound of the hyperparameter range as well as a default value.}
}
\description{
Build an Inducer
}
