% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_temp.R
\name{inducer2Env}
\alias{inducer2Env}
\title{Add new inducer to environment}
\usage{
inducer2Env(.inducer)
}
\arguments{
\item{inducer}{An object of class Inducer.}
}
\value{
nothing, but environment with inducers
}
\description{
Add new inducer to environment
}
