% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerRpart.R
\name{InducerRpart}
\alias{InducerRpart}
\title{Function to create an object of class \code{InducerRpart}}
\usage{
InducerRpart(
  .data = NULL,
  formula,
  weights,
  subset,
  na.action = "na.rpart",
  method,
  model = FALSE,
  x = FALSE,
  y = TRUE,
  parms,
  control,
  cost
)
}
\arguments{
\item{.data}{data object of class \code{Dataset}}

\item{formula}{a formula, with a response but no interaction terms. If this is a
data frame, it is taken as the model frame}

\item{weights}{optional case weights.}

\item{subset}{optional expression saying that only a subset of the rows of the data
should be used in the fit.}

\item{na.action}{the default action deletes all observations for which y is missing,
but keeps those in which one or more predictors are missing.}

\item{method}{one of "anova", "poisson", "class" or "exp". If method is missing then
the routine tries to make an intelligent guess}

\item{model}{if logical: keep a copy of the model frame in the result? If the input
value for model is a model frame}

\item{x}{keep a copy of the x matrix in the result.}

\item{y}{keep a copy of the dependent variable in the result. If missing and model
is supplied this defaults to FALSE.}

\item{parms}{optional parameters for the splitting function.}

\item{control}{a list of options that control details of the rpart algorithm}

\item{cost}{a vector of non-negative costs, one for each variable in the model. Defaults to one for all variables}
}
\value{
a \code{InducerRpart} object
}
\description{
If .data is empty an \code{InducerRpart} object will be created. If .data is
a \code{Dataset} object a rpart model will be fitted.
}
\examples{
inducer <- InducerRpart()
inducer
cars.data <- Dataset(data = cars, target = "dist")
fittedInd <- InducerRpart(.data = cars.data)
fittedInd
}
\seealso{
\code{\link[=fit.InducerRpart]{fit.InducerRpart()}}
}
