% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerXGBoost.R
\name{InducerXGBoost}
\alias{InducerXGBoost}
\title{Function to create an object of class \code{InducerXGBoost}}
\usage{
InducerXGBoost(
  .data = NULL,
  nrounds = 1,
  eta = 0.3,
  gamma = 0,
  max_depth = 6,
  min_child_weight = 1,
  subsample = 1,
  colsample_bytree = 1,
  lambda = 1,
  alpha = 0,
  num_parallel_tree = 1
)
}
\arguments{
\item{.data}{Data object of class \code{Dataset}.}
}
\value{
A \code{InducerXGBoost} object
}
\description{
If .data is empty an \code{InducerXGBoost} object will be created. If .data is a \code{Dataset} object a xgboost model will be fitted
}
\seealso{
\code{\link[=fit.InducerXGBoost]{fit.InducerXGBoost()}}
}
