% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{metainfo.Dataset}
\alias{metainfo.Dataset}
\title{Print the meta Information of a dataset}
\usage{
metainfo.Dataset(
  data,
  targets = data$target,
  type = data$type,
  name = as.name(deparse(substitute(data), 20)[[1]]),
  features = data$features,
  ...
)
}
\arguments{
\item{data}{A dataset}

\item{targets}{The target variable of a dataset}

\item{type}{Regression or Classification}

\item{name}{name}

\item{features}{features}

\item{...}{optional arguments

TO-DO}
}
\value{
prints out the neccessary information
}
\description{
Print the meta Information of a dataset
}
