% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{metainfo}
\alias{metainfo}
\title{S3 method metainfo}
\usage{
metainfo(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Optional arguments passed to or from other methods.}
}
\value{
The metainfo of the object.
}
\description{
Get the information about an object.
}
