% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{modelInfo.Model}
\alias{modelInfo.Model}
\title{modelInfo: print out info of a model}
\usage{
\method{modelInfo}{Model}(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{further arguments}
}
\description{
print the model info of a model
}
\examples{
inducer <- InducerXGBoost()
cars.data <- Dataset(data = cars, target = "dist")
fittedmod <- fit.InducerXGBoost(.inducer = inducer, .data = cars.data)
modelInfo.Model(fittedmod)
}
