% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{modelInfo.Models}
\alias{modelInfo.Models}
\title{modelInfo: get the needed training time in seconds}
\usage{
modelInfo.Models(model, ...)
}
\arguments{
\item{model}{an object of class 'Models'.}
}
\value{
the time needed for the training measured in seconds
}
\description{
This function
}
