% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{modelObject}
\alias{modelObject}
\title{S3 method modelObject}
\usage{
modelObject(model, ...)
}
\arguments{
\item{model}{The model for which the output should be returned.}

\item{...}{Optional arguments passed to or from other methods.}
}
\description{
Get the object generated by the fitted model.
}
