% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_Regression.R
\name{ModelRegression}
\alias{ModelRegression}
\title{Build a ModelRegression object}
\usage{
ModelRegression(data, inducer)
}
\arguments{
\item{name}{description}
}
\description{
Create a ModelRegression object
}
\examples{
# example code
}
