% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{Models}
\alias{Models}
\title{Create a Models object}
\usage{
Models(data, inducer)
}
\arguments{
\item{data}{The data given in an Dataset object.}

\item{inducer}{An Inducer object: The applied inducer}
}
\description{
Create a Models object given a dataset and an inducer.
}
