% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_ModelLm.R
\name{predict.ModelLm}
\alias{predict.ModelLm}
\title{Predict values for \code{fit.InducerLm}}
\usage{
\method{predict}{ModelLm}(model, newdata, ...)
}
\arguments{
\item{model}{a linear model of class \code{ModelLm}}

\item{newdata}{data of class \code{data.frame} or \code{Dataset}}

\item{...}{additional arguments}
}
\value{
An object with the predictions of class \code{numeric} or \code{data.frame}
}
\description{
Predict from the results of a linear model
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
inducer <- InducerLm()
lmfit <- fit.InducerLm(.inducer = inducer, .data = cars.data)
predict.ModelLm(model = lmfit, newdata = data.frame(speed = 10))
predict.ModelLm(model = lmfit, newdata = cars.data[c(1, 2, 3, 4), ])
}
\seealso{
\code{\link[=fit.InducerLm]{fit.InducerLm()}}
}
