% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_ModelRpart.R
\name{predict.ModelRpart}
\alias{predict.ModelRpart}
\title{Predict values for \code{fit.InducerRpart}}
\usage{
\method{predict}{ModelRpart}(object, newdata, ...)
}
\arguments{
\item{object}{a linear model of class \code{ModelRpart}}

\item{newdata}{data of class \code{data.frame} or \code{Dataset}}

\item{...}{additional arguments}
}
\value{
An object with the predictions of class \code{numeric} or \code{data.frame}
}
\description{
Predict from the results of a rpart model
}
\examples{
cars.data <- Dataset(data = cars, target = "dist")
inducer <- InducerLm()
rpartfit <- fit.InducerRpart(.inducer = inducer, .data = cars.data)
predict.ModelRpart(object = rpartfit, newdata = data.frame(speed = 10))
predict.ModelRpart(object = rpartfit, newdata = cars.data[c(1, 2, 3, 4), ])
}
\seealso{
\code{\link[=fit.InducerRpart]{fit.InducerRpart()}}
}
