% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{predict.Models}
\alias{predict.Models}
\title{Predict method for Models}
\usage{
\method{predict}{Models}(model, newdata, ...)
}
\arguments{
\item{model}{A \code{Models} object}

\item{newdata}{A \code{dataset} object for which the values should be fitted}
}
\value{
the fitted values
}
\description{
Predicted values based on a Models object.
}
