% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{predict}
\alias{predict}
\title{S3 method predict}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{A fitted model which should be used for predicting on new data.}

\item{...}{Optional arguments passed to or from other methods.}
}
\value{
A fitted model.
}
\description{
\code{predict} predicts the value of the response variable using a fitted model.
}
