% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{print.Dataset}
\alias{print.Dataset}
\title{A print method for \code{Dataset} objects}
\usage{
\method{print}{Dataset}(x, ...)
}
\arguments{
\item{x}{an object of class 'Dataset'}

\item{...}{other arguments passed to function}
}
\description{
Prints the first and last two rows of \code{data} along with an informative text
on the target and task.
}
