% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_Evaluator.R
\name{print.Evaluator}
\alias{print.Evaluator}
\title{Print an Evaluator}
\usage{
\method{print}{Evaluator}(x, ...)
}
\arguments{
\item{x}{The Evaluator which should be printed.}

\item{...}{optional arguments of the print function.}
}
\description{
Print method for an \code{Evaluator} object.
}
