% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorAccuracyROC.R
\name{print.EvaluatorAccuracy}
\alias{print.EvaluatorAccuracy}
\title{Print an EvaluatorAUC}
\usage{
\method{print}{EvaluatorAccuracy}(x, ...)
}
\description{
Print an \code{EvaluatorAUC} object.
}
\examples{
x <- data.frame(var1 = c(1, 2, 3, 4, 5, 6, 7), target = c(1, 1, 1, 1, 0, 1, 0))
predictions <- c(1)
evaluation <- EvaluatorAccuracy(predictions, x, "target")
print(evaluation)
}
\seealso{
\code{\link[=EvaluatorAccuracy]{EvaluatorAccuracy()}}
}
