% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorAIC.R
\name{print.EvaluatorBIC}
\alias{print.EvaluatorBIC}
\title{Print an EvaluatorBIC}
\usage{
\method{print}{EvaluatorBIC}(x, ...)
}
\arguments{
\item{x}{An \code{EvaluatorBIC} object which should be printed.}

\item{...}{Optional arguments for the print function}
}
\description{
Print an \code{EvaluatorBIC} object.
}
\examples{
cars.ds <- Dataset(data = cars, target = "dist")
mod_fit <- fit(InducerLm(), cars.ds, formula = as.formula("dist~speed"))
evaluation <- EvaluatorBIC(mod_fit)
print(evaluation)
}
