% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorMAE_MSE.R
\name{print.EvaluatorMSE}
\alias{print.EvaluatorMSE}
\title{Print an EvaluatorMSE}
\usage{
\method{print}{EvaluatorMSE}(x, ...)
}
\description{
Print an \code{EvaluatorMSE} object.
}
\examples{
x <- data.frame(var1 = c(1, 1, 1, 1, 0), target = c(1, 2, 3, 4, 5))
predictions <- c(3)
evaluation <- EvaluatorMSE(predictions, x, "target")
print(evaluation)
}
\seealso{
\code{\link[=EvaluatorMSE]{EvaluatorMSE()}}
}
