% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{print.Inducer}
\alias{print.Inducer}
\title{Print method for Inducer object}
\usage{
\method{print}{Inducer}(inducer, ...)
}
\arguments{
\item{inducer}{An inducer being an Inducer object.}
}
\description{
Print an Inducer.
}
