% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerLm.R
\name{print.InducerLm}
\alias{print.InducerLm}
\title{S3 method print for class \code{InducerLm}}
\usage{
\method{print}{InducerLm}(x, ...)
}
\arguments{
\item{x}{object of class \code{InducerLm}.}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
Print an \code{InducerLm} object.
}
\examples{
inducer <- InducerLm()
inducer
}
\seealso{
\code{\link[=InducerLm]{InducerLm()}}
}
