% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerRpart.R
\name{print.InducerRpart}
\alias{print.InducerRpart}
\title{S3 method print for class \code{InducerRpart}}
\usage{
\method{print}{InducerRpart}(x, ...)
}
\arguments{
\item{x}{object of class \code{InducerRpart}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
Print an \code{InducerRpart} object.
}
\examples{
inducer <- InducerRpart()
inducer
}
\seealso{
\code{\link[=InducerRpart]{InducerRpart()}}
}
