% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerXGBoost.R
\name{print.InducerXGBoost}
\alias{print.InducerXGBoost}
\title{S3 method print for class 'InducerXGBoost'}
\usage{
\method{print}{InducerXGBoost}(x, ...)
}
\arguments{
\item{x}{object of class \code{InducerXGBoost}.}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
Print an \code{InducerXGBoost} object.
}
\examples{
inducer <- InducerXGBoost()
print.InducerXGBoost(inducer)
}
\seealso{
\code{\link[=InducerXGBoost]{InducerXGBoost()}}
}
