% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_Classification.R
\name{print.ModelClassification}
\alias{print.ModelClassification}
\title{Printing Classification Models}
\usage{
\method{print}{ModelClassification}(model, ...)
}
\arguments{
\item{model}{object of class \code{ModelClassification}.}
}
\description{
Print a classification model.
}
