% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{print.SplitInstanceCV}
\alias{print.SplitInstanceCV}
\title{Printing a SplitInstanceCV}
\usage{
\method{print}{SplitInstanceCV}(x, ...)
}
\arguments{
\item{x}{Object of class SplitInstanceCV.}

\item{...}{Optional arguments to print methods.}
}
\description{
Print information on the SplitInstanceCV object, including hyperparameters
and some details on the used \code{Dataset}.
}
