% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{register_resampling_strategy}
\alias{register_resampling_strategy}
\title{Register a resampling strategy.}
\usage{
register_resampling_strategy(splt, func, name)
}
\arguments{
\item{splt}{An object of class \code{Split}.}

\item{func}{A function implementing a resampling strategy.}

\item{name}{A string of the symbol used for binding.}
}
\description{
This function just does some basic checks an binds the given name to a
given function in the environment of a concrete \code{Split} object.
}
