% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample}
\alias{resample}
\title{Create a ResamplePrediction}
\usage{
resample(.data, ind, splt)
}
\arguments{
\item{.data}{A \code{Dataset} object.}

\item{ind}{An \code{inducer} instance.}

\item{splt}{A \code{SplitInstance} object.}
}
\description{
This function creates a \code{ResamplePrediction} object that contains
information on resampling procedure applied to given parameters.
}
