% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{set_cv_idx}
\alias{set_cv_idx}
\title{Set indices for k-fold CV with repetitions}
\usage{
set_cv_idx(folds, repeats)
}
\arguments{
\item{folds}{Integer number of folds.}

\item{repeats}{Interger number of repetitions.}
}
\value{
A list of training validation indices
}
\description{
Internal function to set the indices for resampling.
}
