% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{Split}
\alias{Split}
\title{Create a Split object hosting available resampling strategies}
\usage{
Split()
}
\value{
An environment of class \code{Split}.
}
\description{
This function just creates an environment with predefined resampling strategies.
}
