% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{SplitCV}
\alias{SplitCV}
\title{Create a SplitCV object}
\usage{
SplitCV(folds, repeats = 1)
}
\arguments{
\item{folds}{Integer number in [2, Inf)}

\item{repeats}{Integer number in [1, Inf)}
}
\description{
A \code{Split} object implementing the resampling strategy of K-Fold CV.
}
