% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{configuration<-}
\alias{configuration<-}
\title{S3 method configuration<- for class 'InducerLm'}
\usage{
configuration(.inducer) <- value
}
\arguments{
\item{.inducer}{an \code{Inducer} object}

\item{value}{value for changing configuration}
}
\description{
Change configuration of an \code{inducer} object
}
\examples{
inducer <- InducerLm()
inducer
configuration(inducer)$x <- TRUE
}
