% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{configuration.Inducer}
\alias{configuration.Inducer}
\title{S3 method configuration for class 'InducerLm'}
\usage{
\method{configuration}{Inducer}(.inducer, ...)
}
\description{
Get the configuration of an \code{inducer} object
}
