% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{configuration}
\alias{configuration}
\title{Configuration print function for an Inducer object}
\usage{
configuration(inducer)
}
\arguments{
\item{inducer}{An inducer being an Inducer object.}
}
\description{
Print the configuration of an Inducer.
}
