% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{configuration}
\alias{configuration}
\title{S3 method configuration}
\usage{
configuration(...)
}
\arguments{
\item{...}{further arguments}
}
\description{
get configuration of an object
}
\examples{
inducer <- InducerXGBoost()
configuration(inducer)
}
