% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Models.R
\name{configuration}
\alias{configuration}
\title{S3 method configuration}
\usage{
configuration(...)

configuration(...)
}
\arguments{
\item{...}{further arguments}

\item{x}{An object for which the configuration should be returned}
}
\value{
The configuration of the object.
}
\description{
\code{configuration} gets the configuration of a given object.

get configuration of an object
}
\examples{
inducer <- InducerXGBoost()
configuration(inducer)
}
