% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_XGBoost.R
\name{fit.InducerXGBoost}
\alias{fit.InducerXGBoost}
\title{Fit XGBoost model using \code{InducerXGBoost}}
\usage{
\method{fit}{InducerXGBoost}(
  .inducer,
  .data = NULL,
  nrounds = 1,
  eta = 0.3,
  gamma = 0,
  max_depth = 6,
  min_child_weight = 1,
  subsample = 1,
  colsample_bytree = 1,
  lambda = 1,
  alpha = 0,
  num_parallel_tree = 1,
  ...
)
}
\arguments{
\item{.inducer}{A \code{InducerXGBoost} object}

\item{.data}{The data to which the model should be fitted, provided as a \code{Dataset} object.}

\item{nrounds}{number of rounds}

\item{eta}{eta value}

\item{gamma}{gamma}

\item{max_depth}{max depth paramater}

\item{min_child_weight}{min child weight paramater}

\item{subsample}{subsample paramater}

\item{colsample_bytree}{colsample paramater}

\item{lambda}{lambda paramater}

\item{alpha}{alpha paramater}

\item{num_parallel_tree}{number of parallel tree paramater}

\item{...}{further args}
}
\value{
An object of class \code{InducerXGBoost}.
}
\description{
Fit a linear model on the provided data.
}
\examples{
inducer <- InducerXGBoost()
inducer
cars.data <- Dataset(data = cars, target = "dist")
fittedmod <- fit.InducerXGBoost(.inducer = inducer, .data = cars.data)
fittedmod
}
