% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{fit.InducerXGBoost}
\alias{fit.InducerXGBoost}
\title{Fit XGBoost models}
\usage{
fit.InducerXGBoost(.inducer, .data, ...)
}
\arguments{
\item{.inducer}{A \code{Inducer} object}

\item{.data}{A \code{dataset} object for which the values should be fitted}
}
\value{
an xgb object
}
\description{
Uses an Inducer and a dataset to fit a xgb model
}
