% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{fit}
\alias{fit}
\title{S3 method fit}
\usage{
fit(.inducer, ...)
}
\arguments{
\item{.inducer}{An inducer which should be used for the model fitting.}

\item{...}{Optional arguments passed to or from other methods.}
}
\value{
A fitted model.
}
\description{
\code{fit} fits a model on given data.
}
