% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/inducer_temp.R
\name{hyperparameters}
\alias{hyperparameters}
\title{S3 method hyperparameters}
\usage{
hyperparameters(.inducer, ...)

hyperparameters(.inducer, ...)
}
\arguments{
\item{.inducer}{An object of class Inducer.}

\item{...}{Optional arguments.}

\item{x}{An object for which the hyperparameters should be returned.}
}
\value{
The configuration of the object.
}
\description{
\code{hyperparameters} gets the possible hyperparameter values of an object.

Get the Hyperparameters of an inducer.
}
