% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{hyperparameters}
\alias{hyperparameters}
\title{Get the Hyperparameters of an inducer}
\usage{
hyperparameters(inducer, ...)
}
\arguments{
\item{inducer}{An object of class Inducer.}
}
\description{
Get the Hyperparameters of an inducer.
}
