% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{hyperparameters}
\alias{hyperparameters}
\title{S3 method hyperparameters}
\usage{
hyperparameters(x, ...)
}
\arguments{

\item{x}{An object for which the hyperparameters should be returned.}


\item{...}{Optional arguments passed to or from other methods.}
}
\value{
The configuration of the object.
}
\description{
\code{hyperparameters} gets the possible hyperparameter values of an object.
}
