% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{Inducer}
\alias{Inducer}
\title{Create an Inducer}
\usage{
Inducer(name, configuration, hyperparameter)
}
\arguments{
\item{name}{name of the inducer}

\item{configuration}{configuration of the inducer}
}
\description{
Build an Inducer
}
