% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerLm.R
\name{InducerLm}
\alias{InducerLm}
\title{Create an InducerLm object}
\usage{
InducerLm(
  .data = NULL,
  formula,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset
)
}
\arguments{
\item{.data}{An optional data set on which the InducerLm should be fitted}

\item{formula}{An optional object of class "formula" that describes the model to be fitted.}

\item{subset}{An optional object TODO!!!!!}
}
\description{
Create an object of class InducerLm
}
