% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerXGBoost.R
\name{InducerXGBoost}
\alias{InducerXGBoost}
\title{Create an InducerXGBoost}
\usage{
InducerXGBoost(
  .data = NULL,
  nrounds = 1,
  eta = 0.3,
  gamma = 0,
  max_depth = 6,
  min_child_weight = 1,
  subsample = 1,
  colsample_bytree = 1,
  lambda = 1,
  alpha = 0,
  num_parallel_tree = 1
)
}
\description{
Build an InducerXGBoost.
}
