% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer_InducerXGBoost.R
\name{InducerXGBoost}
\alias{InducerXGBoost}
\title{Function to create an object of class \code{InducerXGBoost}}
\usage{
InducerXGBoost(
  .data = NULL,
  nrounds = 1,
  eta = 0.3,
  gamma = 0,
  max_depth = 6,
  min_child_weight = 1,
  subsample = 1,
  colsample_bytree = 1,
  lambda = 1,
  alpha = 0,
  num_parallel_tree = 1
)
}
\arguments{
\item{.data}{Data object of class \code{Dataset}.}

\item{nrounds}{number of rounds}

\item{eta}{eta value}

\item{gamma}{gamma}

\item{max_depth}{max depth paramater}

\item{min_child_weight}{min child weight paramater}

\item{subsample}{subsample paramater}

\item{colsample_bytree}{colsample paramater}

\item{lambda}{lambda paramater}

\item{alpha}{alpha paramater}

\item{num_parallel_tree}{number of parallel tree paramater}
}
\value{
A \code{InducerXGBoost} object
}
\description{
If .data is empty an \code{InducerXGBoost} object will be created. If .data is a \code{Dataset} object
a xgboost model will be fitted
}
\examples{
inducer <- InducerXGBoost()
inducer
cars.data <- Dataset(data = cars, target = "dist")
fittedInd <- InducerXGBoost(.data = cars.data)
fittedInd
}
\seealso{
\code{\link[=fit.InducerXGBoost]{fit.InducerXGBoost()}}
}
