% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlrmini-package}
\alias{mlrmini}
\alias{mlrmini-package}
\title{mlrmini: mlrmini}
\description{
A package for machine learning in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AdvaProgR-2324/mlrmini}
  \item \url{https://advaprogr-2324.github.io/mlrmini/}
  \item Report bugs at \url{https://github.com/AdvaProgR-2324/mlrmini/issues}
}

}
\author{
\strong{Maintainer}: Fake Name \email{teru.mlrmini@gmail.com}

}
