% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{Model}
\alias{Model}
\title{Create a Model object}
\usage{
Model(
  inducer.name,
  inducer.configuration,
  data.name,
  data.target,
  data.features,
  fitted.values = NULL,
  coefficients = NULL,
  modelInfo = NULL,
  model.out,
  model.data
)
}
\arguments{
\item{data}{The data given in an Dataset object.}

\item{inducer}{An Inducer object: The applied inducer}
}
\description{
Create a Model object given a dataset and an inducer.
}
