% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_Classification.R
\name{ModelClassification}
\alias{ModelClassification}
\title{Build a classification model}
\usage{
ModelClassification(data, inducer)
}
\description{
Building a classification model given a Dataset object and an inducer.
}
\examples{
# example code
}
