% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{modelInfo}
\alias{modelInfo}
\title{S3 method modelInfo}
\usage{
modelInfo(x, ...)
}
\arguments{
\item{x}{The model about which the information should be obtained.}

\item{...}{Optional arguments passed to or from other methods.}
}
\value{
The information about the provided \code{Model} object.
}
\description{
Get additional information about a \code{Model} object.
}
