% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{modelObject.Model}
\alias{modelObject.Model}
\title{modelObject: get the print out of a model}
\usage{
\method{modelObject}{Model}(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{further arguments}
}
\description{
print the usual output of a model
}
\examples{
inducer <- InducerXGBoost()
cars.data <- Dataset(data = cars, target = "dist")
fittedmod <- fit.InducerXGBoost(.inducer = inducer, .data = cars.data)
modelObject.Model(fittedmod)
}
