% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Models.R
\name{modelObject}
\alias{modelObject}
\title{S3 method modelObject}
\usage{
modelObject(...)

modelObject(...)
}
\arguments{
\item{...}{further arguments}

\item{x}{The model for which the output should be returned.}
}
\description{
Get the object generated by the fitted model.

get model object
}
\examples{
inducer <- InducerXGBoost()
cars.data <- Dataset(data = cars, target = "dist")
fittedmod <- fit.InducerXGBoost(.inducer = inducer, .data = cars.data)
modelObject.Model(fittedmod)
}
