% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{modelObject}
\alias{modelObject}
\title{S3 method modelObject}
\usage{
modelObject(...)
}
\arguments{
\item{...}{further arguments}
}
\description{
get model object
}
\examples{
inducer <- InducerXGBoost()
cars.data <- Dataset(data = cars, target = "dist")
fittedmod <- fit.InducerXGBoost(.inducer = inducer, .data = cars.data)
modelObject.Model(fittedmod)
}
