% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_EvaluatorMAE_MSE.R
\name{print.EvaluatorMAE}
\alias{print.EvaluatorMAE}
\title{Print an EvaluatorMAE.}
\usage{
\method{print}{EvaluatorMAE}(x, ...)
}

\arguments{
\item{x}{An \code{EvaluatorMAE} object which should be printed.}

\item{...}{Optional arguments for the print function}
}

\description{
Print an \code{EvaluatorMAE} object.
}
\examples{
x <- data.frame(var1 = c(1, 1, 1, 1, 0), target = c(1, 2, 3, 4, 5))
predictions <- c(3)
evaluation <- EvaluatorMAE(predictions, x, "target")
print(evaluation)
}
\seealso{
\code{\link[=EvaluatorMAE]{EvaluatorMAE()}}
}
