% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{print.ModelRegression}
\alias{print.ModelRegression}
\title{Printing Regression Models}
\usage{
\method{print}{ModelRegression}(x, ...)
}
\arguments{
\item{x}{object of class \code{ModelRegression}}

\item{...}{further arguments}
}
\description{
Print a regression model.
}
\examples{
inducer <- InducerXGBoost()
cars.data <- Dataset(data = cars, target = "dist")
fittedmod <- fit.InducerXGBoost(.inducer = inducer, .data = cars.data)
print.ModelRegression(fittedmod)
}
