% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_Regression.R
\name{print.ModelRegression}
\alias{print.ModelRegression}
\title{Printing Regression Models}
\usage{
\method{print}{ModelRegression}(model, ...)
}
\arguments{
\item{model}{object of class \code{ModelRegression}}
}
\description{
Print a regression model.
}
\examples{
# example code
}
