% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{[.Dataset}
\alias{[.Dataset}
\title{Subset a Dataset Object}
\usage{
\method{[}{Dataset}(to_subset, arg_row, arg_col = NULL, ...)
}
\arguments{
\item{to_subset}{A  Dataset object.}

\item{arg_row}{row indices or nothing.}

\item{arg_col}{covariate names or nothing.}

\item{...}{Other arguments passed to function}
}
\value{
A object of type 'Dataset.
}
\description{
This function subsets a custom dataset object based on specified row indices and optional column names.
If column names are not specified, it defaults to using all columns. The function checks if the provided
column names exist in the dataset and whether they include the target variable, which cannot be removed.
}
